refactor: consolidate duplicated GitCommit/get_git_commit_info into util
main.rs carried a copy of serve.rs's git-commit-info fetcher that was
identical except for a hardcoded repo_root = None. One shared
util::get_git_commit_info(hash, repo_root) now serves both. ~70 lines
removed, output JSON unchanged.
